Revert "refactor: ClientTimelineListの表示ロジックを分離"
This reverts commit 2ba23c205b77884e896b530e070c4f95dc4184aa.
